+++ Source/WPF/Tests/Win32.cs
@@ -232,13 +232,13 @@ namespace Walkabout.Tests.Interop
 
         internal static double ConvertToDeviceIndependentPixels(int pixels)
         {
-            return (double)pixels * 96 / (double)DPI;
+            return (double)pixels * 96 / DPI;
         }
 
 
         internal static int ConvertFromDeviceIndependentPixels(double pixels)
         {
-            return (int)(pixels * (double)DPI / 96);
+            return (int)(pixels * DPI / 96);
         }
 
         internal static AutomationElement FindWindow(int processId, string automationId)
